Fix Manifest Test
